+++ src/AlarmClock/Alarm.h
@@ -10,6 +10,7 @@
 #endif
 
 #include "Settings.h"
+#include "Constants.h"
 #include "DaysOfWeek.h"
 #include "MinutesTimeStamp.h"
 #include <RTClib.h> // for datetime